Require two failures to mark server down.
There are three states are server can be in: down, maybe-down, and up.
If a server fails to reply to our 'are you up' test, then we mark it as
being down if it was already marked as being down or maybe down.
Otherwise we flag it as maybe down. If a server passes our test, then we
mark it as being up.